Copy over mpv files on app resume
