+++ app/src/main/java/animiru/feature/mpvfiles/MpvConfig.kt
@@ -1,0 +1,139 @@
+package animiru.feature.mpvfiles
+
+import android.content.Context
+import android.content.res.AssetManager
+import com.hippo.unifile.UniFile
+import eu.kanade.tachiyomi.ui.player.settings.AdvancedPlayerPreferences
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.currentCoroutineContext
+import kotlinx.coroutines.isActive
+import logcat.LogPriority
+import tachiyomi.core.common.util.lang.launchIO
+import tachiyomi.core.common.util.system.logcat
+import tachiyomi.domain.storage.service.StorageManager
+import uy.kohesive.injekt.Injekt
+import uy.kohesive.injekt.api.get
+import java.io.IOException
+import java.io.InputStream
+import java.io.OutputStream
+
+class MpvConfig(
+    private val context: Context,
+    private val storageManager: StorageManager = Injekt.get(),
+    private val advancedPlayerPreferences: AdvancedPlayerPreferences = Injekt.get(),
+) {
+    private val scope = CoroutineScope(Dispatchers.IO)
+    private var copyJob: Job? = null
+
+    fun copyFiles() {
+        if (copyJob?.isActive == true) return
+
+        copyJob = scope.launchIO {
+            val mpvDir = UniFile.fromFile(context.filesDir)!!.createDirectory(MPV_DIR)!!
+            copyUserFiles(mpvDir)
+            copyFontsDirectory(mpvDir)
+            copyAssets(mpvDir)
+        }
+    }
+
+    private suspend fun copyUserFiles(mpvDir: UniFile) {
+        // First, delete all present scripts
+        val scriptsDir = { mpvDir.createDirectory(MPV_SCRIPTS_DIR) }
+        val scriptOptsDir = { mpvDir.createDirectory(MPV_SCRIPTS_OPTS_DIR) }
+        val shadersDir = { mpvDir.createDirectory(MPV_SHADERS_DIR) }
+
+        scriptsDir()?.delete()
+        scriptOptsDir()?.delete()
+        shadersDir()?.delete()
+
+        // Then, copy the user files from the Aniyomi directory
+        if (advancedPlayerPreferences.mpvUserFiles.get()) {
+            storageManager.getScriptsDirectory()?.listFiles()?.forEach { file ->
+                ensureActive()
+                val outFile = scriptsDir()?.createFile(file.name)
+                outFile?.let {
+                    file.openInputStream().copyTo(it.openOutputStream())
+                }
+            }
+            storageManager.getScriptOptsDirectory()?.listFiles()?.forEach { file ->
+                ensureActive()
+                val outFile = scriptOptsDir()?.createFile(file.name)
+                outFile?.let {
+                    file.openInputStream().copyTo(it.openOutputStream())
+                }
+            }
+            storageManager.getShadersDirectory()?.listFiles()?.forEach { file ->
+                ensureActive()
+                val outFile = shadersDir()?.createFile(file.name)
+                outFile?.let {
+                    file.openInputStream().copyTo(it.openOutputStream())
+                }
+            }
+        }
+
+        // Copy over the bridge file
+        val luaFile = scriptsDir()?.createFile("aniyomi.lua")
+        val luaBridge = context.assets.open("aniyomi.lua")
+        luaFile?.openOutputStream()?.bufferedWriter()?.use { scriptLua ->
+            luaBridge.bufferedReader().use { scriptLua.write(it.readText()) }
+        }
+    }
+
+    private suspend fun copyFontsDirectory(mpvDir: UniFile) {
+        // TODO: I think this is a bad hack.
+        //  We need to find a way to let MPV directly access our fonts directory.
+        val fontsDirectory = mpvDir.createDirectory(MPV_FONTS_DIR)!!
+
+        storageManager.getFontsDirectory()?.listFiles()?.forEach { font ->
+            ensureActive()
+            val outFile = fontsDirectory.createFile(font.name)
+            outFile?.let {
+                font.openInputStream().copyTo(it.openOutputStream())
+            }
+        }
+    }
+
+    private fun copyAssets(mpvDir: UniFile) {
+        val assetManager = context.assets
+        val files = arrayOf("subfont.ttf", "cacert.pem")
+        for (filename in files) {
+            var ins: InputStream? = null
+            var out: OutputStream? = null
+            try {
+                ins = assetManager.open(filename, AssetManager.ACCESS_STREAMING)
+                val outFile = mpvDir.createFile(filename)!!
+                // Note that .available() officially returns an *estimated* number of bytes available
+                // this is only true for generic streams, asset streams return the full file size
+                if (outFile.length() == ins.available().toLong()) {
+                    logcat(LogPriority.VERBOSE) { "Skipping copy of asset file (exists same size): $filename" }
+                    continue
+                }
+                out = outFile.openOutputStream()
+                ins.copyTo(out)
+                logcat(LogPriority.WARN) { "Copied asset file: $filename" }
+            } catch (e: IOException) {
+                logcat(LogPriority.ERROR, e) { "Failed to copy asset file: $filename" }
+            } finally {
+                ins?.close()
+                out?.close()
+            }
+        }
+    }
+
+    private suspend inline fun ensureActive() {
+        if (!currentCoroutineContext().isActive) {
+            throw CancellationException()
+        }
+    }
+
+    companion object {
+        const val MPV_DIR = "mpv"
+        const val MPV_FONTS_DIR = "fonts"
+        const val MPV_SCRIPTS_DIR = "scripts"
+        const val MPV_SCRIPTS_OPTS_DIR = "script-opts"
+        const val MPV_SHADERS_DIR = "shaders"
+    }
+}
