+++ app/src/main/java/eu/kanade/tachiyomi/di/AppModule.kt
@@ -3,6 +3,7 @@ package eu.kanade.tachiyomi.di
 import android.app.Application
 import androidx.core.content.ContextCompat
 import androidx.sqlite.driver.bundled.BundledSQLiteDriver
+import animiru.feature.mpvfiles.MpvConfig
 import app.cash.sqldelight.db.SqlDriver
 import com.eygraber.sqldelight.androidx.driver.AndroidxSqliteConfiguration
 import com.eygraber.sqldelight.androidx.driver.AndroidxSqliteDatabaseType
@@ -160,6 +161,7 @@ class AppModule(val app: Application) : InjektModule {
         // <-- AM (SYNC_DRIVE)
 
         // AM -->
+        addSingletonFactory { MpvConfig(app) }
         addSingletonFactory { AudioManager(app) }
         addSingletonFactory { BrightnessManager(app) }
         // <-- AM
