+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/PlayerActivity.kt
@@ -30,7 +30,6 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.content.pm.ActivityInfo
 import android.content.pm.PackageManager
-import android.content.res.AssetManager
 import android.content.res.Configuration
 import android.graphics.Rect
 import android.media.AudioManager
@@ -97,14 +96,10 @@ import tachiyomi.core.common.util.lang.launchNonCancellable
 import tachiyomi.core.common.util.lang.launchUI
 import tachiyomi.core.common.util.lang.withUIContext
 import tachiyomi.core.common.util.system.logcat
-import tachiyomi.domain.storage.service.StorageManager
 import tachiyomi.i18n.MR
 import tachiyomi.i18n.aniyomi.AYMR
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
-import java.io.IOException
-import java.io.InputStream
-import java.io.OutputStream
 import kotlin.math.ceil
 import kotlin.math.floor
 
@@ -123,7 +118,6 @@ class PlayerActivity : BaseActivity() {
     private val audioPreferences: AudioPreferences = Injekt.get()
     private val advancedPlayerPreferences: AdvancedPlayerPreferences = Injekt.get()
     private val subtitlePreferences: SubtitlePreferences = Injekt.get()
-    private val storageManager: StorageManager = Injekt.get()
 
     // AM (DISCORD_RPC) -->
     // private val connectionPreferences: ConnectionPreferences = Injekt.get()
@@ -457,6 +451,7 @@ class PlayerActivity : BaseActivity() {
 
     private fun setupPlayerMPV() {
         val mpvDir = UniFile.fromFile(applicationContext.filesDir)!!.createDirectory(MPV_DIR)!!
+        val fontsDirectory = mpvDir.createDirectory(MPV_FONTS_DIR)!!
 
         val mpvConfFile = mpvDir.createFile("mpv.conf")!!
         advancedPlayerPreferences.mpvConf.get().let { mpvConfFile.writeText(it) }
@@ -464,100 +459,14 @@ class PlayerActivity : BaseActivity() {
         advancedPlayerPreferences.mpvInput.get().let { mpvInputFile.writeText(it) }
 
         player.init(mpv)
-        copyUserFiles(mpvDir)
-        copyAssets(mpvDir)
-        copyFontsDirectory(mpvDir)
 
+        mpv.setPropertyString("sub-fonts-dir", fontsDirectory.filePath!!)
+        mpv.setPropertyString("osd-fonts-dir", fontsDirectory.filePath!!)
         val showBlackBars = if (subtitlePreferences.subtitleBlackBars.get()) "yes" else "no"
         mpv.setOptionString("sub-ass-force-margins", showBlackBars)
         mpv.setOptionString("sub-use-margins", showBlackBars)
         mpv.addLogObserver(playerObserver)
         mpv.addObserver(playerObserver)
-    }
-
-    private fun copyUserFiles(mpvDir: UniFile) {
-        // First, delete all present scripts
-        val scriptsDir = { mpvDir.createDirectory(MPV_SCRIPTS_DIR) }
-        val scriptOptsDir = { mpvDir.createDirectory(MPV_SCRIPTS_OPTS_DIR) }
-        val shadersDir = { mpvDir.createDirectory(MPV_SHADERS_DIR) }
-
-        scriptsDir()?.delete()
-        scriptOptsDir()?.delete()
-        shadersDir()?.delete()
-
-        // Then, copy the user files from the Aniyomi directory
-        if (advancedPlayerPreferences.mpvUserFiles.get()) {
-            storageManager.getScriptsDirectory()?.listFiles()?.forEach { file ->
-                val outFile = scriptsDir()?.createFile(file.name)
-                outFile?.let {
-                    file.openInputStream().copyTo(it.openOutputStream())
-                }
-            }
-            storageManager.getScriptOptsDirectory()?.listFiles()?.forEach { file ->
-                val outFile = scriptOptsDir()?.createFile(file.name)
-                outFile?.let {
-                    file.openInputStream().copyTo(it.openOutputStream())
-                }
-            }
-            storageManager.getShadersDirectory()?.listFiles()?.forEach { file ->
-                val outFile = shadersDir()?.createFile(file.name)
-                outFile?.let {
-                    file.openInputStream().copyTo(it.openOutputStream())
-                }
-            }
-        }
-
-        // Copy over the bridge file
-        val luaFile = scriptsDir()?.createFile("aniyomi.lua")
-        val luaBridge = assets.open("aniyomi.lua")
-        luaFile?.openOutputStream()?.bufferedWriter()?.use { scriptLua ->
-            luaBridge.bufferedReader().use { scriptLua.write(it.readText()) }
-        }
-    }
-
-    private fun copyAssets(mpvDir: UniFile) {
-        val assetManager = this.assets
-        val files = arrayOf("subfont.ttf", "cacert.pem")
-        for (filename in files) {
-            var ins: InputStream? = null
-            var out: OutputStream? = null
-            try {
-                ins = assetManager.open(filename, AssetManager.ACCESS_STREAMING)
-                val outFile = mpvDir.createFile(filename)!!
-                // Note that .available() officially returns an *estimated* number of bytes available
-                // this is only true for generic streams, asset streams return the full file size
-                if (outFile.length() == ins.available().toLong()) {
-                    logcat(LogPriority.VERBOSE) { "Skipping copy of asset file (exists same size): $filename" }
-                    continue
-                }
-                out = outFile.openOutputStream()
-                ins.copyTo(out)
-                logcat(LogPriority.WARN) { "Copied asset file: $filename" }
-            } catch (e: IOException) {
-                logcat(LogPriority.ERROR, e) { "Failed to copy asset file: $filename" }
-            } finally {
-                ins?.close()
-                out?.close()
-            }
-        }
-    }
-
-    private fun copyFontsDirectory(mpvDir: UniFile) {
-        // TODO: I think this is a bad hack.
-        //  We need to find a way to let MPV directly access our fonts directory.
-        lifecycleScope.launchIO {
-            val fontsDirectory = mpvDir.createDirectory(MPV_FONTS_DIR)!!
-
-            storageManager.getFontsDirectory()?.listFiles()?.forEach { font ->
-                val outFile = fontsDirectory.createFile(font.name)
-                outFile?.let {
-                    font.openInputStream().copyTo(it.openOutputStream())
-                }
-            }
-
-            mpv.setPropertyString("sub-fonts-dir", fontsDirectory.filePath!!)
-            mpv.setPropertyString("osd-fonts-dir", fontsDirectory.filePath!!)
-        }
     }
 
     fun setupCustomButtons() {
@@ -574,7 +483,7 @@ class PlayerActivity : BaseActivity() {
             val primaryButtonId = viewModel.primaryButton.value?.id ?: 0L
 
             val customButtonsContent = buildString {
-                append(
+                appendLine(
                     """
                         local lua_modules = mp.find_config_file('scripts')
                         if lua_modules then
@@ -585,7 +494,7 @@ class PlayerActivity : BaseActivity() {
                 )
 
                 buttons.forEach { button ->
-                    append(
+                    appendLine(
                         """
                             ${button.getButtonOnStartup(primaryButtonId)}
                             function button${button.id}()
